Rename file to conform to Kibana naming conventions.
