+++ x-pack/solutions/observability/plugins/observability/public/pages/alert_details/components/alert_details_rule_form_flyout.tsx
